ش

+++ RaftActor/ActorsClasses/HeartbeatClasses..cs
@@ -2,10 +2,10 @@
 {
     public double Id { get; private set; }
     public int Term { get; private set; }
-    public int SenderId { get; private set; }
+    public long SenderId { get; private set; }
     public int LogIndex { get; private set; }
 
-    public Heartbeat(int term, int logIndex, int senderId)
+    public Heartbeat(int term, int logIndex, long senderId)
     {
         Id = (DateTime.UtcNow - new DateTime(1970, 1, 1)).TotalMilliseconds;
         SenderId = senderId;
